+++ R/provider-cortex.R
@@ -64,7 +64,7 @@ chat_cortex_analyst <- function(
   model_spec = NULL,
   model_file = NULL,
   api_args = list(),
-  echo = c("none", "text", "all")
+  echo = c("none", "output", "all")
 ) {
   check_string(account, allow_empty = FALSE)
   check_string(model_spec, allow_empty = FALSE, allow_null = TRUE)