+++ R/provider-openrouter.R
@@ -26,7 +26,7 @@ chat_openrouter <- function(
   model = NULL,
   seed = NULL,
   api_args = list(),
-  echo = c("none", "text", "all")
+  echo = c("none", "output", "all")
 ) {
   turns <- normalize_turns(turns, system_prompt)
   model <- set_default(model, "gpt-4o")